% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceurticles.R
\name{twocolpceurws_article}
\alias{twocolpceurws_article}
\title{CEUR WS 2-colum with page numbers format.}
\usage{
twocolpceurws_article(...)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document()}.}
}
\value{
An R Markdown output format.
}
\description{
CEUR WS 2-colum with page numbers format.
}
\details{
This was adapted from
\url{http://ceur-ws.org/Vol-XXX/samplestyles/}.

\strong{Important note:} This template is _not_ supported by CEUR-WS.
}
\examples{
\dontrun{
library("rmarkdown")
draft("MyArticle.Rmd", template = "twocolpceurws", package = "ceurticles")
render("MyArticle/MyArticle.Rmd")
}

}
